New method: PG::Result#check. (fixes #123)
This method exposes a previously internal-only function for checking the
status of a PG::Result, primarily for people using the async interface.
The check is done for you if you're using the synchronous interface, but
there was no way to get the same behavior for results fetched via
PG::Connection#get_result.

This change also adds the PG::Connection associated with a result as
an instance variable (@connection) on the PG::Result object.

Thanks to @royaltm for the idea.